Убраны конвертации Double->Float. Теперь всё Double.

+++ src/model/GameField.java
@@ -67,15 +67,15 @@ public class GameField implements BallPositionChangedListener {
     public void ballPositionChanged(Ball ball) {
         
         if (ball.getPosition().y < 0) {
-            ball.setPosition(new Point2D.Float(ball.getPosition().x, 0));
+            ball.setPosition(new Point2D.Double(ball.getPosition().x, 0));
             ball.setSpeed(ball.getSpeed().flipVertical());
         }
         
         if (ball.getPosition().x < 0 || ball.getPosition().x + ball.getSize().width > _dimensions.width) {
             if (ball.getPosition().x < 0) {
-                ball.setPosition(new Point2D.Float(0, ball.getPosition().y));
+                ball.setPosition(new Point2D.Double(0, ball.getPosition().y));
             } else {
-                ball.setPosition(new Point2D.Float(_dimensions.width - ball.getSize().width, ball.getPosition().y));
+                ball.setPosition(new Point2D.Double(_dimensions.width - ball.getSize().width, ball.getPosition().y));
             }
             ball.setSpeed(ball.getSpeed().flipHorizontal());
         }
