+++ src/view/ScreenGame.java
@@ -6,7 +6,7 @@ import java.awt.Graphics;
 import java.awt.Graphics2D;
 import java.awt.Image;
 import java.awt.geom.Point2D;
-import java.awt.geom.Point2D.Float;
+import java.awt.geom.Point2D.Double;
 import java.awt.image.BufferedImage;
 
 import model.GameField;
@@ -80,11 +80,11 @@ public class ScreenGame extends GameObject {
 		
 		// Построение уровня
 		// TODO: Загрузка уровня из файла (пока уровень захардкоден)
-		BasicBall newball = new BasicBall(field, new Point2D.Float(40, 160), 8, new Speed2D(0.03, -0.01));
-		BreakableBrick newbrick = new BreakableBrick(field, new Point2D.Float(180, 120), new Dimension(48, 24));
-        BreakableBrick newbrick2 = new BreakableBrick(field, new Point2D.Float(228, 120), new Dimension(48, 24));
-        UnbreakableBrick newbrick3 = new UnbreakableBrick(field, new Point2D.Float(276, 120), new Dimension(48, 24));
-        BasicPaddle paddle = new BasicPaddle(field, new Point2D.Float(0, 584), new Dimension(96, 16));
+		BasicBall newball = new BasicBall(field, new Point2D.Double(40, 160), 8, new Speed2D(0.03, -0.01));
+		BreakableBrick newbrick = new BreakableBrick(field, new Point2D.Double(180, 120), new Dimension(48, 24));
+        BreakableBrick newbrick2 = new BreakableBrick(field, new Point2D.Double(228, 120), new Dimension(48, 24));
+        UnbreakableBrick newbrick3 = new UnbreakableBrick(field, new Point2D.Double(276, 120), new Dimension(48, 24));
+        BasicPaddle paddle = new BasicPaddle(field, new Point2D.Double(0, 584), new Dimension(96, 16));
         
         IngameObjectView ballview = viewfact.newBasicBallView(newball);
         IngameObjectView brick1view = viewfact.newBreakableBrickView(newbrick);
@@ -106,8 +106,8 @@ public class ScreenGame extends GameObject {
         paddle.addBall(newball);
         
         // Тестирование столкновения множества шаров
-        BasicBall ball01 = new BasicBall(field, new Point2D.Float((float) 213.3975, 250), 16, new Speed2D(0.043, -0.025));
-        BasicBall ball02 = new BasicBall(field, new Point2D.Float(400, 200), 16, new Speed2D(-0.05, 0));
+        BasicBall ball01 = new BasicBall(field, new Point2D.Double((float) 213.3975, 250), 16, new Speed2D(0.043, -0.025));
+        BasicBall ball02 = new BasicBall(field, new Point2D.Double(400, 200), 16, new Speed2D(-0.05, 0));
         ball01.addDefaultCollisionBehaviour(BehaviourRebound.getInstance());
         ball02.addDefaultCollisionBehaviour(BehaviourRebound.getInstance());
